change: extract types from model files into seperate modules
