+++ backend/src/types/ClaimDocument.ts
@@ -1,0 +1,9 @@
+import mongoose from "mongoose";
+
+export interface ClaimDocument {
+  claimId: mongoose.Types.ObjectId;
+  description: string;
+  document: string;
+}
+
+export default ClaimDocument;
