+++ backend/src/types/ClaimStatus.ts
@@ -1,0 +1,7 @@
+enum ClaimStatus {
+  SUBMITTED = "SUBMITTED",
+  APPROVED = "APPROVED",
+  REJECTED = "REJECTED",
+}
+
+export default ClaimStatus;
