+++ backend/src/types/Currency.ts
@@ -1,0 +1,7 @@
+enum Currency {
+  BGN = "BGN",
+  EUR = "EUR",
+  GBP = "GBP",
+}
+
+export default Currency;
