+++ backend/src/types/EventType.ts
@@ -1,0 +1,7 @@
+enum EventType {
+  STORM = "STORM",
+  FIRE = "FIRE",
+  //Todo Add other examples
+}
+
+export default EventType;
