+++ backend/src/types/Rigth.ts
@@ -1,0 +1,8 @@
+export enum Right {
+  ADMIN = "ADMIN",
+  CLIENT = "CLIENT",
+  EXPERT = "EXPERT",
+  ACTUARY = "ACTUARY",
+}
+
+export default Right;
